+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/production/turbines/GT_MTE_LargeTurbine_SHSteam.java
@@ -41,7 +41,7 @@ public class GT_MTE_LargeTurbine_SHSteam extends GregtechMetaTileEntity_LargerTu
     }
 
     @Override
-    public int getCasingTextureIndex() {
+    public byte getCasingTextureIndex() {
         return 59;
     }
